oracle js test -> ts
